consolidated charge_stability_dac into charge_stability
