+++ qualibration_graphs/quantum_dots/calibration_utils/charge_stability/parameters.py
@@ -2,10 +2,14 @@ from qua_dashboards.virtual_gates import virtual_layer_adder
 from qualibrate.core import NodeParameters
 from qualibrate.core.parameters import RunnableParameters
 from qualibration_libs.parameters import CommonNodeParameters
-from calibration_utils.common_utils.dc_control_parameters import DCControlParameters
 from calibration_utils.run_video_mode.video_mode_parameters import VideoModeParameters
 
 from typing import List, Literal, Dict, Union, Callable, Optional
+
+
+class DCControlParameters(RunnableParameters):
+    dc_control: bool = False
+    """Apply sweep center via external DC (VirtualDCSet) instead of OPX offset."""
 
 
 class NodeSpecificParameters(RunnableParameters):
@@ -60,7 +64,17 @@ class Parameters(
     VideoModeParameters,
     NodeSpecificParameters,
 ):
-    pass
+    """05a: OPX charge stability map with optional video mode."""
+
+
+class OPXuDACParameters(
+    NodeParameters,
+    CommonNodeParameters,
+    DCControlParameters,
+    NodeSpecificParameters,
+):
+    """05e: OPX charge stability map using user-specific DAC voltage sequences."""
+    plot_pca: bool = True
 
 
 class OPXQDACParameters(
