--- qualibration_graphs/quantum_dots/calibration_utils/charge_stability_dac/parameters.py
@@ -1,176 +1,0 @@
-from qua_dashboards.virtual_gates import virtual_layer_adder
-from qualibrate.core import NodeParameters
-from qualibrate.core.parameters import RunnableParameters
-from qualibration_libs.parameters import CommonNodeParameters
-from calibration_utils.common_utils.dc_control_parameters import DCControlParameters
-
-from typing import List, Literal, Dict, Union, Callable, Optional
-
-
-class NodeSpecificParameters(RunnableParameters):
-    num_shots: int = 10
-    """Number of averages to perform. Default is 100."""
-    sensor_names: List[str] = None
-    """List of sensor dot names to measure in your measurement."""
-    x_axis_name: str = None
-    """The name of the swept element in the X axis."""
-    y_axis_name: str = None
-    """The name of the swept element in the Y axis."""
-    x_points: int = 101
-    """Number of measurement points in the X axis."""
-    y_points: int = 101
-    """Number of measurement points in the Y axis."""
-    x_span: float = 0.01
-    """The X axis span in volts"""
-    y_span: float = 0.01
-    """The Y axis span in volts"""
-    ramp_duration: int = 1000
-    """The ramp duration for the entire sweep."""
-    per_line_compensation: bool = True
-    """Whether to send a compensation pulse at the end of each scan line."""
-    max_compensation_voltage: float = 0.05
-    """The maximum voltage for the compensation pulse."""
-    x_center: Optional[float] = None
-    """The center of the X axis sweep. If dc_control = True, then this will be applied to the external source. Else, it will be applied by the OPX."""
-    y_center: Optional[float] = None
-    """The center of the Y axis sweep. If dc_control = True, then this will be applied to the external source. Else, it will be applied by the OPX."""
-    plot_points: bool = False
-    """Plots the existing points saved in the VirtualGateSet. Default True."""
-    perform_edge_analysis: bool = False
-    """Whether to perform edge analysis on the data."""
-    per_line_wait: int = 0
-    """Wait time at the start of each line, in order to allow the electrostatics to settle."""
-    use_validation: bool = True
-    """Whether to use validation with simulated data."""
-
-
-
-class Parameters(
-    NodeParameters,
-    CommonNodeParameters,
-    DCControlParameters,
-    NodeSpecificParameters,
-):
-    plot_pca: bool = True
-    pixel_hold_duration: int = 1000
-
-
-class OPXQDACParameters(
-    NodeParameters,
-    CommonNodeParameters,
-    DCControlParameters,
-    NodeSpecificParameters,
-):
-    x_from_qdac: bool = False
-    "Check to perform 2D map using the QDAC instead of the OPX"
-    y_from_qdac: bool = False
-    "Check to perform 2D map using the QDAC instead of the OPX"
-    post_trigger_wait_ns: int = 10000
-    """A pause in the QUA program to allow the QDAC to get to the correct level."""
-    qdac_dwell_time_us: float = 200
-    """The dwell time in microseconds for the QDAC."""
-
-
-class SimulationParameters(
-    NodeParameters,
-    CommonNodeParameters,
-    NodeSpecificParameters,
-):
-    pass
-
-
-import numpy as np
-
-
-# Backward-compatible aliases kept for existing nodes/tests that still use the
-# legacy OPX/OPXQDAC parameter class names.
-OPXParameters = Parameters
-# OPXQDACParameters = Parameters
-
-def get_axis_names(node): 
-    """In the case that x_axis_name or y_axis_name is None, assign the first and second elements of QDs."""
-    quantum_dots = list(node.machine.quantum_dots.keys())
-    x_axis_name = node.parameters.x_axis_name
-    y_axis_name = node.parameters.y_axis_name
-
-    if node.parameters.x_axis_name is None:
-        x_axis_name = quantum_dots[0]
-    
-    if node.parameters.y_axis_name is None: 
-        y_axis_name = quantum_dots[1]
-    return x_axis_name, y_axis_name
-
-def get_voltage_arrays(node):
-    """Extract the X and Y voltage arrays from a given node."""
-    x_span, x_center, x_points = node.parameters.x_span, 0, node.parameters.x_points
-    y_span, y_center, y_points = node.parameters.y_span, 0, node.parameters.y_points
-    x_volts, y_volts = np.linspace(
-        x_center - x_span / 2, x_center + x_span / 2, x_points
-    ), np.linspace(y_center - y_span / 2, y_center + y_span / 2, y_points)
-    return x_volts, y_volts
-
-
-from .scan_modes import ScanMode
-
-
-def _find_physical_dc_lists(
-    virtual_dc_set: "VirtualDCSet",
-    axis_name: str,
-    axis_values: List[float],
-) -> Dict[str, Union[List, np.ndarray]]:
-    """Use the VirtualDCSet to yield a dictionary of physical dc_lists to use for the Qdac"""
-
-    full_physical_dicts = {name: [] for name in virtual_dc_set.channels.keys()}
-
-    for value in axis_values:
-        virtual_dict = {axis_name: float(value)}
-        physical_dict = virtual_dc_set.resolve_voltages(virtual_dict)
-
-        for physical_gate in virtual_dc_set.channels.keys():
-            full_physical_dicts[physical_gate].append(physical_dict[physical_gate])
-
-    # Check if the physical list is constant or not
-    physical_lists = {
-        name: arr
-        for name, arr in full_physical_dicts.items()
-        if len(arr) > 1 and not np.allclose(arr, arr[0], atol=1e-8)
-    }
-    return physical_lists
-
-
-def prepare_dc_lists(
-    node,
-    virtual_dc_set_id: str,
-    axis_name: str,
-    axis_values: List[float],
-) -> None:
-    """
-    Prepares the DC list attributes for the QDAC channel. This function assumes the use of the
-    Qdac2 driver from qcodes_contrib_drivers. This also assumes that the VoltageGate objects have
-    their QdacSpec objects configured with the qdac_output_port and opx_trigger_out.
-    """
-    virtual_dc_set = node.machine.virtual_dc_sets[virtual_dc_set_id]
-    physical_dc_lists = _find_physical_dc_lists(virtual_dc_set, axis_name, axis_values)
-
-    trigger = None
-    for ch_name in physical_dc_lists.keys():
-        spec = getattr(virtual_dc_set.channels[ch_name], "qdac_spec", None)
-        if spec is not None:
-            trig = getattr(spec, "qdac_trigger_in", None)
-            if trig is not None:
-                trigger = trig
-                break
-    if trigger is None:
-        raise ValueError(
-            f"No trigger found for the physical outputs associated with the axis {axis_name}"
-        )
-
-    for name, voltages in physical_dc_lists.items():
-        dc_list = node.machine.qdac.channel(
-            virtual_dc_set.channels[name].qdac_spec.qdac_output_port
-        ).dc_list(
-            voltages=voltages,
-            dwell_s=node.parameters.qdac_dwell_time_us / 1e6,
-            stepped=True,
-        )
-        dc_list.start_on_external(trigger=trigger)
