--- qualibration_graphs/quantum_dots/calibration_utils/common_utils/dc_control_parameters.py
@@ -1,6 +1,0 @@
-from qualibrate.core.parameters import RunnableParameters
-
-
-class DCControlParameters(RunnableParameters):
-    dc_control: bool = False
-    """Apply sweep center via external DC (VirtualDCSet) instead of OPX offset."""
